Move some fns from diagnostics -> spectra
also clean up, change some names, enhance docstrings...
