+++ docs/examples/test_schemes_plot.md
@@ -1,16 +1,14 @@
 ---
-kernelspec:
-  display_name: Python 3
-  name: python3
 jupytext:
-  cell_metadata_filter: -all
-  formats: md:myst
-  main_language: python
   text_representation:
     extension: .md
     format_name: myst
     format_version: 0.12
-    jupytext_version: 1.6.0
+    jupytext_version: 1.9.1
+kernelspec:
+  display_name: Python 3
+  language: python
+  name: python3
 ---
 
 # Run each scheme
@@ -19,26 +17,22 @@ We test the following:
 
 * running the model with default case
 * plot methods for {class}`~crt1d.Model` class
-* output xr
-* plots for {class}`~crt1d.Model` class
-* plots for output xr
+* :class:`xarray.Dataset` output
+* plots for output `xarray.Dataset`s
 
-```{code-cell}
+```{code-cell} ipython3
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import xarray as xr
 
 import crt1d as crt
-
-#%matplotlib widget  # currently doesn't work (though the widget buttons show up)
-%matplotlib inline
 ```
 
 ## Run (just once)
 Run the default case (which is loaded automatically when model object is created).
 
-```{code-cell}
+```{code-cell} ipython3
 scheme_names = list(crt.solvers.AVAILABLE_SCHEMES)
 
 ms = []
@@ -55,45 +49,45 @@ dsets = [m.to_xr() for m in ms]
 
 Leaf area profile and leaf angle dist, using the plotting methods attached to the {class}`~crt1d.Model` instance.
 
-```{code-cell}
+```{code-cell} ipython3
 ms[0].plot_canopy()
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 ms[0].plot_leafsoil_spectra()
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 ms[0].plot_toc_spectra()
 ```
 
 ## Plots of results
 
-```{code-cell}
+```{code-cell} ipython3
 crt.diagnostics.plot_compare_band(dsets, band_name="PAR")
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 crt.diagnostics.plot_compare_band(dsets, band_name="solar")
 ```
 
 ### Other diagnostic tools
 
-```{code-cell}
-crt.diagnostics.df_E_balance(dsets).astype(float).round(3)
+```{code-cell} ipython3
+crt.diagnostics.compare_ebal(dsets).astype(float).round(3)
 ```
 
 ## Datasets and their variables
 
-```{code-cell}
+```{code-cell} ipython3
 dsets[0]
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 dsets[0]["zm"]
 ```
 
-```{code-cell}
+```{code-cell} ipython3
 # some have extra outputs
 ms[scheme_names.index("zq")].out_extra
 ```
